Move data classes for v2alpha public API resource keys to a common package.
This allows them to be shared across Duchy and Kingdom services. In particular, this moves them to the JVM package that the API is defined in.
