+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/DataProviderKey.kt
@@ -12,27 +12,24 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.kingdom.service.api.v2alpha
+package org.wfanet.measurement.api.v2alpha
 
 import org.wfanet.measurement.common.ResourceNameParser
 
-private val parser =
-  ResourceNameParser("recurringExchanges/{recurring_exchange}/exchanges/{exchange}")
+private val parser = ResourceNameParser("dataProviders/{data_provider}")
 
-/** [ExchangeKey] of an Exchange. */
-data class ExchangeKey(val recurringExchangeId: String, val exchangeId: String) : ResourceKey {
+/** [DataProviderKey] of a Data Provider. */
+data class DataProviderKey(val dataProviderId: String) : ResourceKey {
   override fun toName(): String {
-    return parser.assembleName(
-      mapOf(IdVariable.RECURRING_EXCHANGE to recurringExchangeId, IdVariable.EXCHANGE to exchangeId)
-    )
+    return parser.assembleName(mapOf(IdVariable.DATA_PROVIDER to dataProviderId))
   }
 
   companion object {
-    val defaultValue = ExchangeKey("", "")
+    val defaultValue = DataProviderKey("")
 
-    fun fromName(resourceName: String): ExchangeKey? {
+    fun fromName(resourceName: String): DataProviderKey? {
       return parser.parseIdVars(resourceName)?.let {
-        ExchangeKey(it.getValue(IdVariable.RECURRING_EXCHANGE), it.getValue(IdVariable.EXCHANGE))
+        DataProviderKey(it.getValue(IdVariable.DATA_PROVIDER))
       }
     }
   }
