+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/ExchangeStepKey.kt
@@ -12,44 +12,40 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.kingdom.service.api.v2alpha
+package org.wfanet.measurement.api.v2alpha
 
 import org.wfanet.measurement.common.ResourceNameParser
 
 private val parser =
   ResourceNameParser(
-    "recurringExchanges/{recurring_exchange}/exchanges/{exchange}/steps/{exchange_step}/" +
-      "attempts/{exchange_step_attempt}"
+    "recurringExchanges/{recurring_exchange}/exchanges/{exchange}/steps/{exchange_step}"
   )
 
-/** [ExchangeStepAttemptKey] of an Exchange Step Attempt. */
-data class ExchangeStepAttemptKey(
+/** [ExchangeStepKey] of an Exchange Step. */
+data class ExchangeStepKey(
   val recurringExchangeId: String,
   val exchangeId: String,
-  val exchangeStepId: String,
-  val exchangeStepAttemptId: String
+  val exchangeStepId: String
 ) : ResourceKey {
   override fun toName(): String {
     return parser.assembleName(
       mapOf(
         IdVariable.RECURRING_EXCHANGE to recurringExchangeId,
         IdVariable.EXCHANGE to exchangeId,
-        IdVariable.EXCHANGE_STEP to exchangeStepId,
-        IdVariable.EXCHANGE_STEP_ATTEMPT to exchangeStepAttemptId
+        IdVariable.EXCHANGE_STEP to exchangeStepId
       )
     )
   }
 
   companion object {
-    val defaultValue = ExchangeStepAttemptKey("", "", "", "")
+    val defaultValue = ExchangeStepKey("", "", "")
 
-    fun fromName(resourceName: String): ExchangeStepAttemptKey? {
+    fun fromName(resourceName: String): ExchangeStepKey? {
       return parser.parseIdVars(resourceName)?.let {
-        ExchangeStepAttemptKey(
+        ExchangeStepKey(
           it.getValue(IdVariable.RECURRING_EXCHANGE),
           it.getValue(IdVariable.EXCHANGE),
-          it.getValue(IdVariable.EXCHANGE_STEP),
-          it.getValue(IdVariable.EXCHANGE_STEP_ATTEMPT)
+          it.getValue(IdVariable.EXCHANGE_STEP)
         )
       }
     }
