+++ src/main/kotlin/org/wfanet/measurement/api/v2alpha/IdVariable.kt
@@ -12,25 +12,25 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.measurement.kingdom.service.api.v2alpha
+package org.wfanet.measurement.api.v2alpha
 
 import org.wfanet.measurement.common.ResourceNameParser
 
-private val parser = ResourceNameParser("modelProviders/{model_provider}")
+internal enum class IdVariable() {
+  RECURRING_EXCHANGE,
+  EXCHANGE,
+  EXCHANGE_STEP,
+  EXCHANGE_STEP_ATTEMPT,
+  DATA_PROVIDER,
+  MODEL_PROVIDER,
+  REQUISITION,
+  PROTOCOL_CONFIG
+}
 
-/** [ModelProviderKey] of a Model Provider. */
-data class ModelProviderKey(val modelProviderId: String) : ResourceKey {
-  override fun toName(): String {
-    return parser.assembleName(mapOf(IdVariable.MODEL_PROVIDER to modelProviderId))
-  }
+internal fun ResourceNameParser.assembleName(idMap: Map<IdVariable, String>): String {
+  return assembleName(idMap.mapKeys { it.key.name.toLowerCase() })
+}
 
-  companion object {
-    val defaultValue = ModelProviderKey("")
-
-    fun fromName(resourceName: String): ModelProviderKey? {
-      return parser.parseIdVars(resourceName)?.let {
-        ModelProviderKey(it.getValue(IdVariable.MODEL_PROVIDER))
-      }
-    }
-  }
+internal fun ResourceNameParser.parseIdVars(resourceName: String): Map<IdVariable, String>? {
+  return parseIdSegments(resourceName)?.mapKeys { IdVariable.valueOf(it.key.toUpperCase()) }
 }
