--- src/main/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha/ResourceKey.kt
@@ -1,20 +1,0 @@
-// Copyright 2021 The Cross-Media Measurement Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package org.wfanet.measurement.duchy.service.api.v2alpha
-
-interface ResourceKey {
-  /** Converts this [ResourceKey] into a resource name. */
-  fun toName(): String
-}
