+++ src/test/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha/BUILD.bazel
@@ -11,9 +11,9 @@ kt_jvm_test(
         "//imports/kotlin/com/nhaarman/mockitokotlin2",
         "//imports/kotlin/kotlin/test",
         "//imports/kotlin/kotlinx/coroutines:core",
+        "//src/main/kotlin/org/wfanet/measurement/api/v2alpha:resource_key",
         "//src/main/kotlin/org/wfanet/measurement/common/grpc/testing",
         "//src/main/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha:requisition_fulfillment_service",
-        "//src/main/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha:resource_key",
         "//src/main/kotlin/org/wfanet/measurement/storage/filesystem:client",
         "//src/main/kotlin/org/wfanet/measurement/storage/testing",
         "//src/main/kotlin/org/wfanet/measurement/system/v1alpha:resource_key",
