+++ src/test/kotlin/org/wfanet/measurement/duchy/service/api/v2alpha/RequisitionFulfillmentServiceTest.kt
@@ -39,6 +39,7 @@ import org.junit.runners.JUnit4
 import org.wfanet.measurement.api.v2alpha.FulfillRequisitionRequest
 import org.wfanet.measurement.api.v2alpha.FulfillRequisitionResponse
 import org.wfanet.measurement.api.v2alpha.Requisition
+import org.wfanet.measurement.api.v2alpha.RequisitionKey
 import org.wfanet.measurement.common.flatten
 import org.wfanet.measurement.common.grpc.testing.GrpcTestServerRule
 import org.wfanet.measurement.common.testing.chainRulesSequentially
